Updated Expander Pi Library
Several updates to the Expander Pi library.  ABE_helper is now merged
with the main library.  The DAC now supports setting the gain to 1 or 2.
The RTC now has support for writing to the onboard SRAM.  Various other
small bug fixes and changes to the demo scripts to work with the updated
library.